requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22.0
neo4j>=5.0
openai>=1.40.0
pydantic>=1.10
python-dotenv>=1.0.0
